Add --keep-logs to Util.py

+++ scripts/Util.py
@@ -2053,8 +2053,11 @@ class LocalProcessController(ProcessController):
 
         def teardown(self, current, success):
             if self.traceFile:
-                if success or current.driver.isInterrupted():
-                    os.remove(self.traceFile)
+                if not current.driver.keepLogs and (success or current.driver.isInterrupted()):
+                    try:
+                        os.remove(self.traceFile)
+                    except FileNotFoundError:
+                        pass
                 else:
                     current.writeln("saved {0}".format(self.traceFile))
 
@@ -2873,7 +2876,7 @@ class Driver:
     @classmethod
     def getSupportedArgs(self):
         return ("dlrR", ["debug", "driver=", "filter=", "rfilter=", "host=", "host-ipv6=", "host-bt=", "interface=",
-                         "controller-app", "valgrind", "languages=", "rlanguages="])
+                         "controller-app", "valgrind", "languages=", "rlanguages=", "keep-logs"])
 
     @classmethod
     def usage(self):
@@ -2895,6 +2898,7 @@ class Driver:
         print("--interface=<IP>      The multicast interface to use to discover controllers.")
         print("--controller-app      Start the process controller application.")
         print("--valgrind            Start executables with valgrind.")
+        print("--keep-logs           Keep log files from successful runs.")
 
     def __init__(self, options, component):
         self.component = component
@@ -2910,6 +2914,7 @@ class Driver:
         self.languages = [self.languages] if self.languages else []
         self.rlanguages = []
         self.failures = []
+        self.keepLogs = False
 
         parseOptions(self, options, { "d": "debug",
                                       "r" : "filters",
@@ -2918,7 +2923,8 @@ class Driver:
                                       "rfilter" : "rfilters",
                                       "host-ipv6" : "hostIPv6",
                                       "host-bt" : "hostBT",
-                                      "controller-app" : "controllerApp"})
+                                      "controller-app" : "controllerApp",
+                                      "keep-logs" : "keepLogs"})
 
         if self.languages:
             self.languages = [i for sublist in [l.split(",") for l in self.languages] for i in sublist]
